Use new quantifications endpoints
